Made a collision detection system.

+++ src/Entities/Entity.cpp
@@ -2,10 +2,10 @@
 
 Entity::Entity() {}
 
-Entity::Entity(size_t id) : m_id(id) {}
+Entity::Entity(int id) : m_id(id) {}
 
 
-const size_t Entity::getId() const
+const int Entity::getId() const
 {
   return m_id;
 }
