+++ src/Entities/Entity.h
@@ -4,11 +4,11 @@
 
 class Entity
 {
-  size_t m_id = 0;
+  int m_id = 0;
 
 public:
   Entity();
-  Entity(size_t id);
+  Entity(int id);
 
-  const size_t getId() const;
+  const int getId() const;
 };
