+++ src/Managers/ConfigurationManager.cpp
@@ -7,13 +7,13 @@ ConfigurationManager& ConfigurationManager::Instance()
 }
 
 
-const std::vector<size_t>& ConfigurationManager::getConfigs(const std::string& entity) const
+const std::vector<int>& ConfigurationManager::getConfigs(const std::string& entity) const
 {
   return m_configMap.at(entity);
 }
 
 
-void ConfigurationManager::setConfigs(std::map<std::string, std::vector<size_t>>& entityConfigMap)
+void ConfigurationManager::setConfigs(std::map<std::string, std::vector<int>>& entityConfigMap)
 {
   m_configMap = entityConfigMap;
 }
